components folder structure edited
